Removed Addressables
-removed: AddressablesManager.cs
-removed: Addressables package
-added: AssetManager.cs
