--- Assets/AddressableAssetsData.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3a11222986ef4664aaaaa21227edb986
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
