--- Assets/AddressableAssetsData/AddressableAssetSettings.asset
@@ -1,98 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 468a46d0ae32c3544b7d98094e6448a9, type: 3}
-  m_Name: AddressableAssetSettings
-  m_EditorClassIdentifier: 
-  m_DefaultGroup: 22e21baebde909944b82cf2e299f4b72
-  m_BuildRemoteCatalog: 0
-  m_DisableCatalogUpdateOnStart: 0
-  m_UniqueBundleIds: 0
-  m_RemoteCatalogBuildPath:
-    m_Id: 
-  m_RemoteCatalogLoadPath:
-    m_Id: 
-  m_overridePlayerVersion: 
-  m_GroupAssets:
-  - {fileID: 11400000, guid: 9357306a84550134c956ad1cd2127e02, type: 2}
-  - {fileID: 11400000, guid: 6b09e032e11759d4da224fbfef4b97db, type: 2}
-  - {fileID: 11400000, guid: 565fe525040576a4b83df8905e1304b3, type: 2}
-  - {fileID: 11400000, guid: e68b0084790c4d2489fab00f854e26aa, type: 2}
-  - {fileID: 11400000, guid: e82e85ff7dc32a645b0c1672f41a57e8, type: 2}
-  - {fileID: 11400000, guid: 852afba9ca7e33541a83786e8522d273, type: 2}
-  - {fileID: 11400000, guid: 42e10871950c7454ea5417eaac9672f5, type: 2}
-  - {fileID: 11400000, guid: 37c868c203bc8cb49ab9abc1565d77d3, type: 2}
-  - {fileID: 11400000, guid: bc964ce18584ad946891d7c85d2ee191, type: 2}
-  m_BuildSettings:
-    m_CompileScriptsInVirtualMode: 0
-    m_CleanupStreamingAssetsAfterBuilds: 1
-    m_LogResourceManagerExceptions: 1
-    m_BundleBuildPath: Temp/com.unity.addressables/AssetBundles
-  m_ProfileSettings:
-    m_Profiles:
-    - m_InheritedParent: 
-      m_Id: 32db550ba9bf4b0469eddcb3fe0e0e78
-      m_ProfileName: Default
-      m_Values:
-      - m_Id: 91e3994d42ca4fc4b935540f08df0256
-        m_Value: '[UnityEditor.EditorUserBuildSettings.activeBuildTarget]'
-      - m_Id: 67f9e5029a5c23842abb1348855148fb
-        m_Value: '[UnityEngine.AddressableAssets.Addressables.BuildPath]/[BuildTarget]'
-      - m_Id: de9a67307f19e864ca764a9ffc470475
-        m_Value: '{UnityEngine.AddressableAssets.Addressables.RuntimePath}/[BuildTarget]'
-      - m_Id: 8f1704f11dbc28d4d94c1faa78cf063e
-        m_Value: ServerData/[BuildTarget]
-      - m_Id: 31ef157523224594e9807fb069a42680
-        m_Value: http://localhost/[BuildTarget]
-    m_ProfileEntryNames:
-    - m_Id: 91e3994d42ca4fc4b935540f08df0256
-      m_Name: BuildTarget
-      m_InlineUsage: 0
-    - m_Id: 67f9e5029a5c23842abb1348855148fb
-      m_Name: LocalBuildPath
-      m_InlineUsage: 0
-    - m_Id: de9a67307f19e864ca764a9ffc470475
-      m_Name: LocalLoadPath
-      m_InlineUsage: 0
-    - m_Id: 8f1704f11dbc28d4d94c1faa78cf063e
-      m_Name: RemoteBuildPath
-      m_InlineUsage: 0
-    - m_Id: 31ef157523224594e9807fb069a42680
-      m_Name: RemoteLoadPath
-      m_InlineUsage: 0
-    m_ProfileVersion: 1
-  m_LabelTable:
-    m_LabelNames:
-    - Prefabs
-    - Items
-    - Scenery
-    - Structures
-    - Characters
-    - Furniture
-  m_SchemaTemplates: []
-  m_GroupTemplateObjects:
-  - {fileID: 11400000, guid: 201e5b8141f86f948951f4a9973de7a9, type: 2}
-  m_InitializationObjects: []
-  m_CertificateHandlerType:
-    m_AssemblyName: 
-    m_ClassName: 
-  m_ActivePlayerDataBuilderIndex: 3
-  m_DataBuilders:
-  - {fileID: 11400000, guid: 628f3aa30315dae40bb5c801eebc7abf, type: 2}
-  - {fileID: 11400000, guid: 3503549f4b9244140bbc9e217e203bc9, type: 2}
-  - {fileID: 11400000, guid: 016f26b18707c7e4b8f8758a26344d5b, type: 2}
-  - {fileID: 11400000, guid: ab838e01fe2ddf643974dda4e4a5e705, type: 2}
-  m_ActiveProfileId: 32db550ba9bf4b0469eddcb3fe0e0e78
-  m_HostingServicesManager:
-    m_HostingServiceInfos: []
-    m_Settings: {fileID: 11400000}
-    m_NextInstanceId: 0
-    m_RegisteredServiceTypeRefs: []
