--- Assets/AddressableAssetsData/AddressableAssetSettings.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 18c507c874153ae4eba3c65b9e3d999e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
