--- Assets/AddressableAssetsData/AssetGroupTemplates.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8090847760b40ba4cae13bfdbbaa1e31
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
