--- Assets/AddressableAssetsData/AssetGroupTemplates/Packed Assets.asset
@@ -1,67 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &-1590293986138156161
-MonoBehaviour:
-  m_ObjectHideFlags: 1
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5834b5087d578d24c926ce20cd31e6d6, type: 3}
-  m_Name: ContentUpdateGroupSchema
-  m_EditorClassIdentifier: 
-  m_Group: {fileID: 0}
-  m_StaticContent: 0
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1a3c5d64ac83548c09dd1678b9f6f1cd, type: 3}
-  m_Name: Packed Assets
-  m_EditorClassIdentifier: 
-  m_SchemaObjects:
-  - {fileID: 7725322772522896684}
-  - {fileID: -1590293986138156161}
-  m_Description: Pack assets into asset bundles.
-  m_Settings: {fileID: 0}
---- !u!114 &7725322772522896684
-MonoBehaviour:
-  m_ObjectHideFlags: 1
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e5d17a21594effb4e9591490b009e7aa, type: 3}
-  m_Name: BundledAssetGroupSchema
-  m_EditorClassIdentifier: 
-  m_Group: {fileID: 0}
-  m_Compression: 1
-  m_IncludeInBuild: 1
-  m_BundledAssetProviderType:
-    m_AssemblyName: 
-    m_ClassName: 
-  m_ForceUniqueProvider: 0
-  m_UseAssetBundleCache: 1
-  m_UseAssetBundleCrc: 1
-  m_Timeout: 0
-  m_ChunkedTransfer: 0
-  m_RedirectLimit: -1
-  m_RetryCount: 0
-  m_BuildPath:
-    m_Id: 
-  m_LoadPath:
-    m_Id: 
-  m_BundleMode: 0
-  m_AssetBundleProviderType:
-    m_AssemblyName: 
-    m_ClassName: 
-  m_BundleNaming: 0
