--- Assets/AddressableAssetsData/AssetGroupTemplates/Packed Assets.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 201e5b8141f86f948951f4a9973de7a9
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
