--- Assets/AddressableAssetsData/AssetGroups.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f4e894c62dc654c48b401706ae00a4b4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
