--- Assets/AddressableAssetsData/AssetGroups/Built In Data.asset
@@ -1,34 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbb281ee3bf0b054c82ac2347e9e782c, type: 3}
-  m_Name: Built In Data
-  m_EditorClassIdentifier: 
-  m_GroupName: Built In Data
-  m_Data:
-    m_SerializedData: []
-  m_GUID: 5ecb7444faa8dae46bb684d7bc3da5af
-  m_SerializeEntries:
-  - m_GUID: EditorSceneList
-    m_Address: EditorSceneList
-    m_ReadOnly: 1
-    m_SerializedLabels: []
-    m_mainAssetType: 
-  - m_GUID: Resources
-    m_Address: Resources
-    m_ReadOnly: 1
-    m_SerializedLabels: []
-    m_mainAssetType: 
-  m_ReadOnly: 1
-  m_Settings: {fileID: 11400000, guid: 18c507c874153ae4eba3c65b9e3d999e, type: 2}
-  m_SchemaSet:
-    m_Schemas:
-    - {fileID: 11400000, guid: de7cbafee8b7b1146951ee19a6f02173, type: 2}
