--- Assets/AddressableAssetsData/AssetGroups/Built In Data.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9357306a84550134c956ad1cd2127e02
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
