--- Assets/AddressableAssetsData/AssetGroups/Common.asset
@@ -1,42 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbb281ee3bf0b054c82ac2347e9e782c, type: 3}
-  m_Name: Common
-  m_EditorClassIdentifier: 
-  m_GroupName: Common
-  m_Data:
-    m_SerializedData: []
-  m_GUID: 22e21baebde909944b82cf2e299f4b72
-  m_SerializeEntries:
-  - m_GUID: 578562fb1a22e104aa03fc3450a6ea08
-    m_Address: prefab_item
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - default
-    - Default
-    - Prefabs
-    m_mainAssetType: UnityEngine.GameObject, UnityEngine.CoreModule, Version=0.0.0.0,
-      Culture=neutral, PublicKeyToken=null
-  - m_GUID: 7c2f46a6c01686946891a3b27aef6494
-    m_Address: texture_Log
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - default
-    - Items
-    m_mainAssetType: UnityEngine.Texture2D, UnityEngine.CoreModule, Version=0.0.0.0,
-      Culture=neutral, PublicKeyToken=null
-  m_ReadOnly: 0
-  m_Settings: {fileID: 11400000, guid: 18c507c874153ae4eba3c65b9e3d999e, type: 2}
-  m_SchemaSet:
-    m_Schemas:
-    - {fileID: 11400000, guid: ba2a080e5b61af040b809b74b7f67c59, type: 2}
-    - {fileID: 11400000, guid: c7fa972d08ed36f42b075caed868d2b3, type: 2}
