--- Assets/AddressableAssetsData/AssetGroups/Common.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6b09e032e11759d4da224fbfef4b97db
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
