--- Assets/AddressableAssetsData/AssetGroups/Schemas.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4475757ad7f39454889f2ec978fe6ce9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
