--- Assets/AddressableAssetsData/AssetGroups/Schemas/22e21baebde909944b82cf2e299f4b72_BundledAssetGroupSchema.asset
@@ -1,36 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e5d17a21594effb4e9591490b009e7aa, type: 3}
-  m_Name: 22e21baebde909944b82cf2e299f4b72_BundledAssetGroupSchema
-  m_EditorClassIdentifier: 
-  m_Group: {fileID: 11400000, guid: 6b09e032e11759d4da224fbfef4b97db, type: 2}
-  m_Compression: 1
-  m_IncludeInBuild: 1
-  m_BundledAssetProviderType:
-    m_AssemblyName: Unity.ResourceManager, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null
-    m_ClassName: UnityEngine.ResourceManagement.ResourceProviders.BundledAssetProvider
-  m_ForceUniqueProvider: 0
-  m_UseAssetBundleCache: 1
-  m_UseAssetBundleCrc: 1
-  m_Timeout: 0
-  m_ChunkedTransfer: 0
-  m_RedirectLimit: -1
-  m_RetryCount: 0
-  m_BuildPath:
-    m_Id: 67f9e5029a5c23842abb1348855148fb
-  m_LoadPath:
-    m_Id: de9a67307f19e864ca764a9ffc470475
-  m_BundleMode: 0
-  m_AssetBundleProviderType:
-    m_AssemblyName: Unity.ResourceManager, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null
-    m_ClassName: UnityEngine.ResourceManagement.ResourceProviders.AssetBundleProvider
-  m_BundleNaming: 0
