--- Assets/AddressableAssetsData/AssetGroups/Schemas/22e21baebde909944b82cf2e299f4b72_BundledAssetGroupSchema.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c7fa972d08ed36f42b075caed868d2b3
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
