--- Assets/AddressableAssetsData/AssetGroups/Schemas/22e21baebde909944b82cf2e299f4b72_ContentUpdateGroupSchema.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ba2a080e5b61af040b809b74b7f67c59
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
