--- Assets/AddressableAssetsData/AssetGroups/Schemas/5ecb7444faa8dae46bb684d7bc3da5af_PlayerDataGroupSchema.asset
@@ -1,17 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b1487f5d688e4f94f828f879d599dbdc, type: 3}
-  m_Name: 5ecb7444faa8dae46bb684d7bc3da5af_PlayerDataGroupSchema
-  m_EditorClassIdentifier: 
-  m_Group: {fileID: 0}
-  m_IncludeResourcesFolders: 1
-  m_IncludeBuildSettingsScenes: 1
