--- Assets/AddressableAssetsData/AssetGroups/Schemas/5ecb7444faa8dae46bb684d7bc3da5af_PlayerDataGroupSchema.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: de7cbafee8b7b1146951ee19a6f02173
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
