--- Assets/AddressableAssetsData/DataBuilders.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 24477b448dda2844ebf4fc88818766d3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
