--- Assets/AddressableAssetsData/DataBuilders/BuildScriptFastMode.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 628f3aa30315dae40bb5c801eebc7abf
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
