--- Assets/AddressableAssetsData/DataBuilders/BuildScriptPackedMode.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ab838e01fe2ddf643974dda4e4a5e705
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
