--- Assets/AddressableAssetsData/DataBuilders/BuildScriptPackedPlayMode.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 016f26b18707c7e4b8f8758a26344d5b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
