--- Assets/AddressableAssetsData/DataBuilders/BuildScriptVirtualMode.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3503549f4b9244140bbc9e217e203bc9
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
