--- Assets/AddressableAssetsData/DefaultObject.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4d1fb5e93ebe4c24c951750889ae0c80
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
