--- Assets/AddressableAssetsData/Windows.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9e425f97f918b1542b074d9c22b33f60
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
