--- Assets/AddressableAssetsData/Windows/addressables_content_state.bin.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4b485f26e0e91994396337354fac15cb
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
