+++ Assets/Scenes/Main.unity
@@ -14396,15 +14396,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 1756808113}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c08f20fa9148ccd48a4969a74b42d4f2, type: 3}
+  m_Script: {fileID: 11500000, guid: 65ff6165a826bdb4ea12a41ecf139d43, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  prefabs_common:
-  - m_AssetGUID: 578562fb1a22e104aa03fc3450a6ea08
-    m_SubObjectName: 
-  textures_common:
-  - m_AssetGUID: 7c2f46a6c01686946891a3b27aef6494
-    m_SubObjectName: Log
 --- !u!1 &1759515093
 GameObject:
   m_ObjectHideFlags: 0
